Build non-fingerprinted CSS file copies on deploy
When static pages share the same CSS as the Ember application, we don't
have a straightforward way to ensure that they load the same CSS without
a flicker or delay. broccoli-asset-rev fingerprints our assets on each
deploy, making it difficult for the static pages to know which CSS file
to download.

One solution is to use broccoli's assetMap, but this solution requires
JavaScript to load CSS, which makes our pages that use the
implementation look incorrect as the CSS loads after JavaScript files
and other assets have loaded.

The solution here is cribbed from the Ruby `sprockets-redirect` gem. On
deploy, we make a non-fingerprinted copy of each CSS file, allowing
permitted pages to request the latest CSS with a consistent canonical
name without knowledge of the fingerprint.